+++ lib/Memory/memory_mram.h
@@ -12,6 +12,7 @@
  *
  * The memory array is of 4 Mbit.
  *
+ *
  * #### Because it is SPI based, the pins are the following:
  *    Supply voltage,
  *    Ground,
@@ -21,6 +22,7 @@
  *    not(WriteProtect),
  *    Serial Input,
  *    Serial Output
+ *
  *
  * #### SPI configuration:
  *
@@ -42,7 +44,8 @@
  *  SLEEP = Enter sleep mode
  *  WAKE = Exit sleep mode
  *
- * How to write and read:
+ *
+ * #### How to write and read:
  * 
  * Chip select must be 1 before a write command. It has to be first
  * set to 0, and then a WREN (Write Enable) command must be first executed.
@@ -61,6 +64,7 @@
  *    call any other instruction sequence before RDSR after a READ, or simply
  *    call RDSR twice, so that the second execution is able to output correctly.
  *    This is a restriction of this MRAM, probably due to technical problems.
+ *
  *
  * #### There is a 1 byte status register:
  * SRWD - 0 - 0 - 0 - BP1 - BP2 - WEL - 0
@@ -93,6 +97,7 @@
  * 1   |  1   | 1  |    Protected      |    Unprotected      |    Unprotected
  * Depending on BP1, BP0, the protection varies.
  *
+ *
  * #### Instructions
  *
  * An instruction is made up of an OPCODE followed by address or dummy bytes
@@ -124,6 +129,11 @@
  * 
  * Check the previously mentioned datasheet for a great explanation on the
  * sequences for each instruction.
+ * 
+ * I assume there is only one SPI for all the memories, so that the clock,
+ *  input, output lines are all the same for the different memories, and
+ *  because of that, a single SPI.begin() on the sketch will setup those
+ *  lines for all the memories to use.
  */
 
 #pragma once
@@ -133,11 +143,6 @@
 
 // Pins
 #define CHIP_SELECT_MRAM 3
-
-// I assume there is only one SPI for all the memories, so that the clock,
-// input, output lines are all the same for the different memories, and
-// because of that, a single SPI.begin() on the sketch will setup those
-// lines for all the memories to use.
 
 // opcodes
 #define WREN_MRAM 6
@@ -191,7 +196,7 @@ public:
    *  8 MBit.
    * @pre 0 <= address <= 2^20 - 1
    */
-  uint8_t readByte(uint32_t address);
+  uint8_t readByte(size_t address);
 
   /**
    * @brief Read N consecutive bytes by incrementing an initialAddress
@@ -204,12 +209,12 @@ public:
    * @param size amount of bytes to read.
    * @pre 0 <= initialAddress <= 2^20 - 1
    */
-  void readNBytes(uint32_t initialAddress, uint8_t* buffer, int size);
+  void readNBytes(size_t initialAddress, uint8_t* buffer, int size);
 
   /**
    * @brief Write a byte.
    * 
-   * Note: write needs to be enabled for the instruction to take effect.
+   * NOTE: write needs to be enabled for the instruction to take effect.
    * 
    * Also, if power goes down, the last incompleted byte to write will be lost.
    *
@@ -220,18 +225,18 @@ public:
    * @pre Write is enabled
    * @pre Region to write at is not protected.
    */
-  void writeByte(uint8_t byteToWrite, uint32_t address);
+  void writeByte(uint8_t byteToWrite, size_t address);
 
   /**
    * @brief Write N consecutive bytes by incrementing an initialAddress
    *    N times.
    * 
-   * Note: due to the bus being most significant first, then write most
+   * NOTE: due to the bus being most significant first, then write most
    *    significant first.
    * 
    * Also, if power goes down, the last incompleted byte to write will be lost.
    * 
-   * Note: write needs to be enabled for the instruction to take effect.
+   * NOTE: write needs to be enabled for the instruction to take effect.
    * 
    * @param buffer bytes that will substitute the old bytes in memory.
    * @param size amount of bytes to write.
@@ -242,11 +247,11 @@ public:
    * @pre Write is enabled
    * @pre Region to write at is not protected.
    */
-  void writeNBytes(uint8_t* buffer, int size, uint32_t initialAddress);
+  void writeNBytes(uint8_t* buffer, int size, size_t initialAddress);
 
 private:
   // because readByte, readNBytes, writeByte, writeNBytes are similar and will
   // likely stay similar. So this is a auxiliary function for them.
-  void transferNBytes(uint8_t opcode, uint32_t address, uint8_t* buffer,
+  void transferNBytes(uint8_t opcode, size_t address, uint8_t* buffer,
       int amountOfBytes);
 };
